requirement.txt

#scikit-learn==1.0.2
pandas==1.4.4
scikit-optimize==0.9.0
matplotlib==3.6.3
seaborn==0.12.2
tabulate==0.9.0
# scipy==1.9.1
tqdm==4.64.1
joblib~=1.0.0
ax-platform==0.2.10
torch
torchvision
tensorboard
tensorboardX
tensorflow
ax-platform==0.2.10
psutil==5.9.4
scikit-image
mpmath==1.3.0
patsy==0.5.3
umap-learn==0.5.3
shapely==2.0.0
openpyxl==3.0.10
xgboost==1.0.0
neptune
fastapi==0.89.1
mlflow[extras]==2.2.1
threadpoolctl==3.1.0
transformers
opencv-python
mysql-connector-python
